requirements.txt
python-telegram-bot==20.3
spotipy==2.23.0
tzdata==2023.3
